requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
cryptography==41.0.7
fastapi==0.115.0
uvicorn==0.30.6
redis==5.0.1
